+++ apps/extension/src/main.tsx
@@ -2,13 +2,14 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { ReduxProvider } from "@providers/Redux";
 import { StyleProvider } from "@providers/Style";
-import { App } from "@app";
+import { RouterProvider } from "react-router";
+import { routers } from "./Router";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <ReduxProvider>
       <StyleProvider>
-        <App />
+        <RouterProvider router={routers} />
       </StyleProvider>
     </ReduxProvider>
   </StrictMode>