+++ apps/extension/src/models/ApplicationSettingsModel.ts
@@ -4,7 +4,7 @@ export type Language = keyof typeof Languages;
 export type Theme = "light" | "dark";
 
 export interface ApplicationSettingsModel {
-  theme?: Theme;
+  theme: Theme;
   useSystemTheme: boolean;
   language: Language;
   showIPV6: boolean;