+++ apps/extension/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [2.0.0] - 03-07-2025
 
+- ✨ feat: add routing(home and settings)
+- ✨ feat: add functionality for dark/light/system theme
 - ✨ feat: add redux/redux-persist and initial state for settings
 - 🔧 chore: add Chackra UI
 - 🔧 chore: add manifest.json for chrome extension API